% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmetr.R
\docType{package}
\name{bmetr-package}
\alias{bmetr-package}
\title{\pkg{bmetr}: Beethoven's Metronome Analysis}
\description{
Supporting data and methods for Martín-Castro & Ucar (2020). Includes
functions and utilities for modelling a mechanical metronome and filtering
tempo signals. See the vignettes for further details.
}
\references{
Martín-Castro A, Ucar I (2020).
Conductors' tempo choices shed light over Beethoven's metronome.
\emph{PLOS ONE} 15(12), e0243616. \doi{10.1371/journal.pone.0243616}
}
\author{
Almudena Martín-Castro, Iñaki Ucar
}
