% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{errors}
\alias{errors}
\alias{gather_errors}
\alias{unite_errors}
\alias{separate_errors}
\title{Data Wrangling with \pkg{errors}}
\usage{
gather_errors(.data, key, value, ...)

unite_errors(.data, suffix = ".se")

separate_errors(.data, suffix = ".se")
}
\arguments{
\item{.data}{a data frame.}

\item{key, value}{names of the new key and value columns, as strings.}

\item{...}{a selection of columns to gather.}

\item{suffix}{suffix for the error column(s).}
}
\description{
Auxiliary functions to perform some common data wrangling operations in an
\pkg{errors}-aware fashion, i.e., to preserve uncertainty metadata.
}
\details{
\code{gather_errors} is the \pkg{errors}-aware equivalent to
\code{tidyr::gather}.

\code{unite_errors} and \code{separate_errors} are similar to
\code{tidyr::unite} and \code{tidyr::separate}, but only for \code{errors}
objects. For example, for each variable \code{var} of class \code{errors},
\code{separate_errors} will store the numeric value in \code{var} and will
create another variable called \code{var.se} for the errors. Similarly, for
each pair of variables \code{var} and \code{var.se}, \code{unite_errors} will
store an \code{errors} object in \code{var} and will remove \code{var.se}.
Note that the suffix \code{.se} can be changed.
}
