% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_peaks}
\alias{find_peaks}
\title{Find Local Peaks}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{numeric vector.}

\item{m}{size of the neighbourhood.}
}
\value{
A vector of indices.
}
\description{
Find local maxima in a neighbourhood. To find local minima, just provide
\code{-x} instead of \code{x}.
}
\references{
Based on \url{https://github.com/stas-g/findPeaks}.
}
