% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metr.data}
\alias{metr.data}
\alias{metr.neewer}
\alias{metr.marks}
\alias{metr.params}
\title{Metronomes Data}
\format{
\bold{\code{metr.neewer}} provides oscillation frequency measurements for the
\code{Neewer} metronome.
\tabular{ll}{
  \code{mark}    \tab metronome mark, in pulses per minute. \cr
  \code{bpm}     \tab measured frequency, in pulses per minute. \cr
  \code{bpm.se}  \tab standard error of \code{bpm}, in the same units.
}

\bold{\code{metr.marks}} provides position measurements of each metronome
mark for all the metronomes.
\tabular{ll}{
  \code{model}   \tab metronome model. \cr
  \code{mark}    \tab metronome mark, in pulses per minute. \cr
  \code{r}       \tab distance from the shaft to the mark, in mm. \cr
  \code{r.se}    \tab standard error of \code{r}, in the same units.
}

\bold{\code{metr.params}} provides measurements of fixed dimensions for all
the metronomes. All the additional columns suffixed with \code{.se} are the
standard errors of the corresponding variables in the same units.
\tabular{ll}{
  \code{model}   \tab metronome model. \cr
  \code{rcm}     \tab position of the center of mass of the upper mass with
                      respect to the metronome mark, in mm. \cr
  \code{l}       \tab length of the rod over the shaft, in mm. \cr
  \code{R}       \tab distance from the shaft to the center of mass of the
                      lower mass, in mm. \cr
  \code{L}       \tab length of the rod below the shaft, in mm. \cr
  \code{A}       \tab oscillation amplitude, in degrees.
}
}
\usage{
metr.neewer

metr.marks

metr.params
}
\description{
Measurements of five models of metronome: \code{Neewer}, \code{Neewer (photo)},
\code{Patent}, \code{TB 06} and \code{TB 07}. The first two correspond to a
contemporary metronome, a Neewer NW-707. The difference is that \code{Neewer}
corresponds to precise direct measurements in a dismantled metronome, while
\code{Neewer (photo)} measurements were taken from a photograph. The rest of
the models follow this methodology. The third one, \code{Patent}, corresponds
to the original diagram from Maelzel's 1815 English patent. Finally,
\code{TB 06} and \code{TB 07} are metronomes number 6 and 7, respectively,
from Tony Bingham's collection (see the references).
}
\references{
Bingham, T. and Turner, A. (2017) \emph{Metronomes and Musical
Time: Catalogue of the Tony Bingham Collection at the exhibition AUF TAKT!
Held in the Museum für Musik, Basel}. London. ISBN: 978-0-946113-11-8.
}
\keyword{datasets}
