% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{metr_fit}
\alias{metr_fit}
\alias{metr_predict}
\title{Metronome Model Fit}
\usage{
metr_fit(.data, params, by = "model")

metr_predict(.fit, by = "model")
}
\arguments{
\item{.data}{measurements of distance to the shaft for each metronome mark.
It must contain the variables \code{model}, \code{mark} and \code{r}, as,
e.g., the \code{\link{metr.marks}} data set.}

\item{params}{the rest of the dimensions: correction of the center of mass
of the upper mass \code{rcm}, length of the rod above the shaft \code{l},
length of the rod below the shaft \code{L}, distance of the lower mass to
the shaft \code{R} and oscillation amplitude \code{A}, as, e.g., the
\code{\link{metr.params}} data set.}

\item{by}{variable to group by; e.g., the metronome model, to fit
several metronomes.}

\item{.fit}{output from \code{metr_fit}.}
}
\value{
A grouped data frame with \code{model}, \code{M.} (and error),
\code{mu.} (and error) and the \code{fit} object.
}
\description{
Fit metronome's nondimensionalized masses, \code{M.} and \code{m.} (see
\code{\link{metr_model}}), from a set of measurements and parameters.
}
