% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{metr_model}
\alias{metr_model}
\alias{metr_model_base}
\alias{metr_model_angle}
\alias{metr_model_friction}
\alias{metr_model_bias}
\alias{metr_model_r}
\title{Metronome Model}
\usage{
metr_model(r, R = 50, M. = 5, l = 220, L = R, mu. = 0, g = 9.807,
  A = 0, ep0 = 0)

metr_model_base(r, R = 50, M. = 5, l = 220, L = R, mu. = 0, g = 9.807)

metr_model_angle(A, terms = 8)

metr_model_friction(r, R = 50, M. = 5, l = 220, L = R, mu. = 0, ep0 = 0)

metr_model_bias(mark, R = 50, M. = 5, l = 220, L = R, mu. = 0,
  g = 9.807, A = 0, ep0 = 0, shift = 0)

metr_model_r(mark, R = 50, M. = 5, l = 220, L = R, mu. = 0, g = 9.807, A = 0)
}
\arguments{
\item{r}{distance from the shaft to the center of mass of the upper mass, in mm.}

\item{R}{distance from the shaft to the center of mass of the lower mass, in mm.}

\item{M.}{nondimensionalized lower mass (lower mass divided by upper mass).}

\item{l}{length of the rod above the shaft, in mm.}

\item{L}{length of the rod below the shaft, in mm.}

\item{mu.}{nondimensionalized rod mass (rod mass divided by upper mass).}

\item{g}{standard gravity, in m/s^2.}

\item{A}{oscillation amplitude, in degrees.}

\item{ep0}{nondimensional friction parameter (\code{>= 0}).}

\item{terms}{number of terms to approximate the angle correction.}

\item{mark}{metronome mark, in pulses per minute.}

\item{shift}{include a shift of the rod with respect to the scale, in mm.}
}
\value{
\code{metr_model}, \code{metr_model_base} and \code{metr_model_bias}
return the oscillation frequency in pulses per minute.

\code{metr_model_angle} and \code{metr_model_friction} return a value \code{>= 1}.

\code{metr_model_r} returns the distance from the shaft to the upper mass in mm.
}
\description{
Functions to compute a metronome's oscillation frequency. \code{metr_model}
is the main function. It computes the base value (\code{metr_model_base}),
which includes the contribution of the mass of the rod (\code{mu.}), and
applies corrections due to the oscillation angle (\code{metr_model_angle})
and the friction (\code{metr_model_friction}).
}
\details{
Function \code{metr_model_bias} computes, for a given metronome mark, the
resulting oscillation frequency for an alteration of any parameter. To this
end, it uses \code{metr_model_r} to first calculate the position of the upper
mass for this metronome mark.

A value of \code{mu.=0} means that the rod is considered massless.
A value of \code{A=0} means that no correction is applied due to the
oscillation amplitude. A value of \code{ep0=0} means that the movement is
frictionless. A value of \code{ep0=0.5} means that the metronome stops due
to friction at the lowest possible frequency.

To compute the frequency bias for any parameter, two values must be supplied
to \code{metr_model_bias}. For example, if the original position of the lower
mass is \code{50} mm, and we want to calculate the result of lowering it down
\code{5} mm, then \code{R=c(50, 55)} should be supplied.
}
