% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sym.data}
\alias{sym.data}
\alias{sym.marks}
\alias{sym.recordings}
\alias{sym.duration}
\alias{sym.window}
\alias{sym.sample}
\title{Beethoven's Symphonies Data}
\format{
\bold{\code{sym.marks}} provides complete timing information extracted from
Beethoven's scores: character annotations, tempo markings, beats and bars for
the nine symphonies. Note that some movements have several character
annotations (more than one section).
\tabular{ll}{
  \code{symphony}  \tab symphony number. \cr
  \code{movement}  \tab movement number. \cr
  \code{character} \tab character annotation in this section. \cr
  \code{mark}      \tab metronome mark. \cr
  \code{beat}      \tab beats per bar. \cr
  \code{bar.tot}   \tab total number of bars in this section. \cr
  \code{bar.rep}   \tab total number of repeated bars in this section. \cr
  \code{bar.1}     \tab total number of bars in first repetition boxes. \cr
  \code{tsig}      \tab time signature, as expected by
                        \code{\link{tmp_rectify_tsig}} \cr
  \code{section}   \tab convenience section identifier (one per character).
}
Therefore, the theoretical duration, in minutes, \emph{without repetitions}
for each section can be computed as follows:
\code{(bar.tot - bar.1) * beat / mark}.
And the theoretical duration, in minutes, \emph{with repetitions} is just the
previous value \code{+ bar.rep * beat / mark}.

\bold{\code{sym.recordings}} provides complete information about the albums
analyzed in this work (the nine symphonies for 36 different conductors).
\tabular{ll}{
  \code{conductor} \tab conductor's name. \cr
  \code{orchestra} \tab orchestra(s)'s name. \cr
  \code{title}     \tab album title. \cr
  \code{label}     \tab album label. \cr
  \code{upc}       \tab Universal Product Code. \cr
  \code{date}      \tab recording dates. \cr
  \code{year}      \tab release date. \cr
  \code{ptype}     \tab performance type: romantic, Historically Informed
                        (HI) or under HI influence.
}

\bold{\code{sym.duration}} provides performed duration (track length) for the
nine symphonies and 36 different conductors.
\tabular{ll}{
  \code{symphony}  \tab symphony number. \cr
  \code{movement}  \tab movement number. \cr
  \code{conductor} \tab conductor's name. \cr
  \code{duration}  \tab track length, in seconds.
}

\bold{\code{sym.window}} provides continuous tempo measurements per symphony,
movement and conductor by means of a sliding window.
\tabular{ll}{
  \code{symphony}  \tab symphony number. \cr
  \code{movement}  \tab movement number. \cr
  \code{conductor} \tab conductor's name. \cr
  \code{n}         \tab sample index. Samples with \code{n=0} correspond to
                        an estimation for the entire track. \cr
  \code{start}     \tab start time for the sliding window, in seconds. \cr
  \code{duration}  \tab window length, in seconds. \cr
  \code{tempo}     \tab estimated tempo.
}

\bold{\code{sym.sample}}, in contrast to \code{sym.window}, provides a single
tempo estimation for each symphony, movement and conductor. The sample was
collected at the end of each track (the "coda" of the movement), where the
tempo is arguably more stable. It has the same variables as \code{sym.window},
except for \code{n}. Instead, \code{sym.sample} contains a \code{section}
column to identify the section from which the tempo was sampled. This section
identifier corresponds to the homonymous one present in \code{sym.marks}.
}
\usage{
sym.marks

sym.recordings

sym.duration

sym.window

sym.sample
}
\description{
Six data sets with tempo measurements plus additional information about
Beethoven's nine symphonies.
\itemize{
  \item Data sets \code{sym.marks}, \code{sym.recordings} and \code{sym.duration}
  provide detailed information about Beethoven's annotations as well as the
  recordings selected for this study (the complete symphonies from 36 different
  conductors).
  \item Data sets \code{sym.window}, and \code{sym.sample}
  provide tempo measurements extracted using the Marsyas framework (and the
  tempo estimation algorithm by Percival and Tzanetakis; see the references)
  and distinct methodologies.
}
}
\references{
Tzanetakis, G. and Cook, P. (2000) "MARSYAS: a framework for audio analysis."
\emph{Organised Sound}, 4(3):169-175

Percival, G. and Tzanetakis, G. (2013) "An effective, simple tempo estimation
method based on self-similarity and regularity." \emph{IEEE International
Conference on Acoustics, Speech and Signal Processing}, 241-245.

Percival, G. and Tzanetakis, G. (2014) "Streamlined Tempo Estimation Based on
Autocorrelation and Cross-correlation with Pulses." \emph{IEEE/ACM Trans.
Audio, Speech and Lang. Proc.}, 22(12):1765-1776.
}
\keyword{datasets}
